refactor: implement file metadata retrieval by ID, update upload endpoint, and enhance exception handling
