+++ src/main/java/com/nexaworks/rafiq/service/authentication/AuthServiceImpl.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.service.authentication.implementaion;
+package com.nexaworks.rafiq.service.authentication;
 
 import java.time.Instant;
 import java.util.UUID;
@@ -15,8 +15,6 @@ import com.nexaworks.rafiq.entities.Token;
 import com.nexaworks.rafiq.entities.User;
 import com.nexaworks.rafiq.exception.custom.TokenInvalidException;
 import com.nexaworks.rafiq.repository.UserRepository;
-import com.nexaworks.rafiq.service.authentication.AuthService;
-import com.nexaworks.rafiq.service.authentication.JwtService;
 import com.nexaworks.rafiq.service.user.TokenService;
 import com.nexaworks.rafiq.utils.AuthSessionManager;
