+++ src/main/java/com/nexaworks/rafiq/service/authentication/OAuth2GoogleService.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.service.authentication.implementaion;
+package com.nexaworks.rafiq.service.authentication;
 
 import java.io.IOException;
 import java.security.GeneralSecurityException;
@@ -13,7 +13,6 @@ import com.google.api.client.googleapis.auth.oauth2.GoogleIdTokenVerifier;
 import com.nexaworks.rafiq.dto.response.auth.LoginResponse;
 import com.nexaworks.rafiq.entities.User;
 import com.nexaworks.rafiq.exception.custom.GoogleAuthException;
-import com.nexaworks.rafiq.service.authentication.OAuth2Service;
 import com.nexaworks.rafiq.service.user.UserService;
 import com.nexaworks.rafiq.utils.AuthSessionManager;
