+++ src/main/java/com/nexaworks/rafiq/service/labReports/LabTestServiceImpl.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.service.labReports.implementation;
+package com.nexaworks.rafiq.service.labReports;
 
 import java.io.IOException;
 import java.time.Instant;
@@ -8,8 +8,6 @@ import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 
-import com.nexaworks.rafiq.service.labReports.LabResultService;
-import com.nexaworks.rafiq.service.labReports.LabTestService;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
