+++ src/main/java/com/nexaworks/rafiq/service/medicine/GroupServiceImpl.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.service.medicine.implementation;
+package com.nexaworks.rafiq.service.medicine;
 
 import java.util.Optional;
 import java.util.UUID;
@@ -19,7 +19,6 @@ import com.nexaworks.rafiq.exception.custom.GroupIsAlreadyExistsException;
 import com.nexaworks.rafiq.exception.custom.GroupNotFoundException;
 import com.nexaworks.rafiq.exception.custom.MedicineNotFound;
 import com.nexaworks.rafiq.repository.GroupRepository;
-import com.nexaworks.rafiq.service.medicine.GroupService;
 import com.nexaworks.rafiq.service.patient.PatientService;
 
 import lombok.RequiredArgsConstructor;
