+++ src/main/java/com/nexaworks/rafiq/service/notification/EmailSenderServiceImpl.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.service.notification.implementation;
+package com.nexaworks.rafiq.service.notification;
 
 import java.util.Map;
 
@@ -13,7 +13,6 @@ import org.thymeleaf.context.Context;
 import org.thymeleaf.spring6.SpringTemplateEngine;
 
 import com.nexaworks.rafiq.exception.custom.MailSenderException;
-import com.nexaworks.rafiq.service.notification.EmailSenderService;
 
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
