+++ src/main/java/com/nexaworks/rafiq/service/patient/WeightHistoryServiceImpl.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.service.patient.implementation;
+package com.nexaworks.rafiq.service.patient;
 
 import java.time.LocalDate;
 
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Service;
 import com.nexaworks.rafiq.entities.Patient;
 import com.nexaworks.rafiq.entities.WeightHistory;
 import com.nexaworks.rafiq.repository.WeightHistoryRepository;
-import com.nexaworks.rafiq.service.patient.WeightHistoryService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
