+++ src/main/java/com/nexaworks/rafiq/service/user/PasswordServiceImpl.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.service.user.implementation;
+package com.nexaworks.rafiq.service.user;
 
 import java.time.Instant;
 import java.util.Optional;
@@ -16,8 +16,6 @@ import com.nexaworks.rafiq.entities.Token;
 import com.nexaworks.rafiq.entities.User;
 import com.nexaworks.rafiq.exception.custom.TokenInvalidException;
 import com.nexaworks.rafiq.repository.UserRepository;
-import com.nexaworks.rafiq.service.user.PasswordService;
-import com.nexaworks.rafiq.service.user.TokenService;
 import com.nexaworks.rafiq.service.authentication.AuthService;
 
 import lombok.RequiredArgsConstructor;
