+++ src/test/java/com/nexaworks/rafiq/unit/service/AuthServiceImplTest.java
@@ -29,8 +29,8 @@ import com.nexaworks.rafiq.entities.Token;
 import com.nexaworks.rafiq.entities.User;
 import com.nexaworks.rafiq.exception.custom.TokenInvalidException;
 import com.nexaworks.rafiq.repository.UserRepository;
+import com.nexaworks.rafiq.service.authentication.AuthServiceImpl;
 import com.nexaworks.rafiq.service.authentication.JwtService;
-import com.nexaworks.rafiq.service.authentication.implementaion.AuthServiceImpl;
 import com.nexaworks.rafiq.service.user.TokenService;
 import com.nexaworks.rafiq.utils.AuthSessionManager;
