+++ src/test/java/com/nexaworks/rafiq/unit/service/EmailSenderServiceImplTest.java
@@ -20,7 +20,7 @@ import org.thymeleaf.spring6.SpringTemplateEngine;
 
 import com.nexaworks.rafiq.exception.custom.MailSenderException;
 import com.nexaworks.rafiq.service.notification.EmailSenderService;
-import com.nexaworks.rafiq.service.notification.implementation.EmailSenderServiceImpl;
+import com.nexaworks.rafiq.service.notification.EmailSenderServiceImpl;
 
 import jakarta.mail.MessagingException;
 import jakarta.mail.Session;
