+++ src/test/java/com/nexaworks/rafiq/unit/service/LabTestServiceImplTest.java
@@ -22,10 +22,10 @@ import com.nexaworks.rafiq.exception.custom.LabTestException;
 import com.nexaworks.rafiq.repository.LabTestRepository;
 import com.nexaworks.rafiq.repository.PatientRepository;
 import com.nexaworks.rafiq.service.file.ImageService;
-import com.nexaworks.rafiq.service.labReports.implementation.LabResultServiceImpl;
-import com.nexaworks.rafiq.service.labReports.implementation.LabTestServiceImpl;
-import com.nexaworks.rafiq.service.patient.implementation.PatientServiceImpl;
-import com.nexaworks.rafiq.service.user.implementation.UserServiceImpl;
+import com.nexaworks.rafiq.service.labReports.LabResultServiceImpl;
+import com.nexaworks.rafiq.service.labReports.LabTestServiceImpl;
+import com.nexaworks.rafiq.service.patient.PatientServiceImpl;
+import com.nexaworks.rafiq.service.user.UserServiceImpl;
 
 public class LabTestServiceImplTest {
     @Mock
