+++ src/test/java/com/nexaworks/rafiq/unit/service/MedicineServiceImplTest.java
@@ -28,9 +28,9 @@ import com.nexaworks.rafiq.exception.custom.MedicineAlreadyExist;
 import com.nexaworks.rafiq.exception.custom.MedicineLimit;
 import com.nexaworks.rafiq.repository.MedicineRepository;
 import com.nexaworks.rafiq.secheduler.service.QuartzSchedulerService;
-import com.nexaworks.rafiq.service.medicine.implementation.DrugServiceImpl;
-import com.nexaworks.rafiq.service.medicine.implementation.GroupServiceImpl;
-import com.nexaworks.rafiq.service.medicine.implementation.MedicineServiceImpl;
+import com.nexaworks.rafiq.service.medicine.DrugServiceImpl;
+import com.nexaworks.rafiq.service.medicine.GroupServiceImpl;
+import com.nexaworks.rafiq.service.medicine.MedicineServiceImpl;
 import com.nexaworks.rafiq.service.patient.PatientService;
 import com.nexaworks.rafiq.service.user.UserService;
