+++ src/test/java/com/nexaworks/rafiq/unit/service/PasswordServiceImplTest.java
@@ -31,8 +31,8 @@ import com.nexaworks.rafiq.entities.User;
 import com.nexaworks.rafiq.exception.custom.TokenInvalidException;
 import com.nexaworks.rafiq.repository.UserRepository;
 import com.nexaworks.rafiq.service.authentication.AuthService;
+import com.nexaworks.rafiq.service.user.PasswordServiceImpl;
 import com.nexaworks.rafiq.service.user.TokenService;
-import com.nexaworks.rafiq.service.user.implementation.PasswordServiceImpl;
 
 @ExtendWith(MockitoExtension.class)
 @DisplayName("PasswordServiceImpl Unit Tests")
