+++ src/test/java/com/nexaworks/rafiq/unit/service/PatientServiceImplTest.java
@@ -10,7 +10,7 @@ import org.mockito.MockitoAnnotations;
 
 import com.nexaworks.rafiq.entities.User;
 import com.nexaworks.rafiq.repository.PatientRepository;
-import com.nexaworks.rafiq.service.patient.implementation.PatientServiceImpl;
+import com.nexaworks.rafiq.service.patient.PatientServiceImpl;
 
 @DisplayName("PatientService Test Cases")
 public class PatientServiceImplTest {
