+++ src/test/java/com/nexaworks/rafiq/unit/service/PdfExtractorServiceImplTest.java
@@ -23,10 +23,10 @@ import org.springframework.mock.web.MockMultipartFile;
 import com.itextpdf.text.DocumentException;
 import com.nexaworks.rafiq.entities.User;
 import com.nexaworks.rafiq.exception.custom.EmptyFileException;
-import com.nexaworks.rafiq.service.ai.implementation.GeminiService;
+import com.nexaworks.rafiq.service.ai.GeminiService;
 import com.nexaworks.rafiq.service.authentication.AuthService;
-import com.nexaworks.rafiq.service.file.implementation.PdfExtractorServiceImpl;
-import com.nexaworks.rafiq.service.labReports.implementation.LabTestServiceImpl;
+import com.nexaworks.rafiq.service.file.PdfExtractorServiceImpl;
+import com.nexaworks.rafiq.service.labReports.LabTestServiceImpl;
 
 public class PdfExtractorServiceImplTest {
     @Mock
