+++ src/test/java/com/nexaworks/rafiq/unit/service/UserServiceImplTest.java
@@ -33,13 +33,13 @@ import com.nexaworks.rafiq.exception.custom.RegistrationException;
 import com.nexaworks.rafiq.exception.custom.TokenInvalidException;
 import com.nexaworks.rafiq.exception.custom.TokenNotFoundException;
 import com.nexaworks.rafiq.repository.UserRepository;
-import com.nexaworks.rafiq.service.doctor.implementation.DoctorServiceImpl;
-import com.nexaworks.rafiq.service.doctor.implementation.SpecializationServiceImpl;
+import com.nexaworks.rafiq.service.doctor.DoctorServiceImpl;
+import com.nexaworks.rafiq.service.doctor.SpecializationServiceImpl;
 import com.nexaworks.rafiq.service.file.ImageService;
-import com.nexaworks.rafiq.service.patient.implementation.PatientServiceImpl;
-import com.nexaworks.rafiq.service.user.implementation.RoleServiceImpl;
-import com.nexaworks.rafiq.service.user.implementation.TokenServiceImpl;
-import com.nexaworks.rafiq.service.user.implementation.UserServiceImpl;
+import com.nexaworks.rafiq.service.patient.PatientServiceImpl;
+import com.nexaworks.rafiq.service.user.RoleServiceImpl;
+import com.nexaworks.rafiq.service.user.TokenServiceImpl;
+import com.nexaworks.rafiq.service.user.UserServiceImpl;
 import com.nexaworks.rafiq.utils.AuthSessionManager;
 
 import jakarta.servlet.http.HttpServletResponse;
